[#34] refactor: 중복되는 부분과 상수화 시켜야 된다고 판단되는 부분들을 분류
