+++ frontend/src/components/calendar/Calendar.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
+import { DAYS } from '../../constants/date';
+
 import CalendarHeader from './CalendarHeader';
 import CalendarDays from './CalendarDays';
 import CalendarCells from './CalendarCells';
 import useDate from '../../hooks/useDate';
-
-const DAYS = ['일', '월', '화', '수', '목', '금', '토'];
 
 function Calendar() {
 	const { date: currentDate, prevMonth, resetDate, nextMonth } = useDate();
