+++ frontend/src/constants/color.js
@@ -1,0 +1,3 @@
+const COLORS = ['red', 'green', 'black', 'pink', 'blue', 'orange', 'gray'];
+
+export { COLORS };
