+++ frontend/src/hooks/useSchedule.js
@@ -1,17 +1,8 @@
 import { useEffect } from 'react';
 import useInput from './useInput';
-import format from 'date-fns/format';
+
 import { addDays } from 'date-fns';
-
-const getFormatDate = date => {
-	return format(date, 'yyyy-MM-dd');
-};
-
-const getEndTime = startTime => {
-	const [nowHour] = startTime.split(':');
-	const hour = Number(nowHour) + 1;
-	return nowHour < '23' ? `${hour < 10 ? '0' + hour : hour}:00` : `00:00`;
-};
+import { getFormatDate, getEndTime } from '../util/date';
 
 function useValidateSchedule({
 	initialTitle,
